Revert "Add localization_channel unit test"
This reverts commit fd3997d1f9bbcb59756bbfdbe5b6626aca5c4390.
