+++ shell/platform/tizen/channels/localization_channel_stub.cc
@@ -6,23 +6,10 @@
 
 namespace flutter {
 
-void LocalizationChannel::SendPlatformResolvedLocale() {}
+LocalizationChannel::LocalizationChannel(FlutterTizenEngine* engine) {}
 
-std::vector<FlutterLocale*> LocalizationChannel::GetFlutterLocales() {
-  FlutterLocale* flutter_locale = GetFlutterLocale(nullptr);
-  std::vector<FlutterLocale*> flutter_locales;
-  flutter_locales.push_back(flutter_locale);
-  return flutter_locales;
-}
+LocalizationChannel::~LocalizationChannel() {}
 
-FlutterLocale* LocalizationChannel::GetFlutterLocale(const char* locale) {
-  FlutterLocale* flutter_locale = new FlutterLocale;
-  flutter_locale->struct_size = sizeof(FlutterLocale);
-  flutter_locale->language_code = new char[1];
-  flutter_locale->country_code = new char[1];
-  flutter_locale->script_code = new char[1];
-  flutter_locale->variant_code = new char[1];
-  return flutter_locale;
-}
+void LocalizationChannel::SendLocales() {}
 
 }  // namespace flutter
